For Python3.9

+++ ambari-agent/src/main/python/ambari_agent/alerts/script_alert.py
@@ -54,7 +54,7 @@ class ScriptAlert(BaseAlert):
     self.parameters = {}
 
     # will force a kinit even if klist says there are valid tickets (4 hour default)
-    self.kinit_timeout = long(config.get('agent', 'alert_kinit_timeout', BaseAlert._DEFAULT_KINIT_TIMEOUT))
+    self.kinit_timeout = int(config.get('agent', 'alert_kinit_timeout', BaseAlert._DEFAULT_KINIT_TIMEOUT))
 
     if 'path' in alert_source_meta:
       self.path = alert_source_meta['path']
